Make the cache size configurable
