+++ configs/sample.conf
@@ -56,6 +56,11 @@ com.socrata.soql-server-pg {
   }
 
   instance = ${secondary-instance}
+
+  cache {
+    max-cached-results: 1000,
+    max-result-size: 1000
+  }
 }
 
 com.socrata.coordinator.secondary-watcher = {
