+++ soql-server-pg/docker/Dockerfile
@@ -23,6 +23,8 @@ ENV PG_SECONDARY_INSTANCE pg
 ENV PG_SECONDARY_WORK_MEM 256MB
 ENV PG_SECONDARY_CLIENT_MSG_LEVEL error
 ENV PG_SECONDARY_MAX_POOL_SIZE 40
+ENV PG_SECONDARY_MAX_CACHED_RESULTS 10000
+ENV PG_SECONDARY_MAX_CACHED_RESULT_SIZE 10240
 ENV JAVA_MAX_METASPACE 64m
 ENV JAVA_OOM_BEHAVIOR -XX:+ExitOnOutOfMemoryError
 ENV MAX_CONCURRENT_REQUESTS_PER_DATASET 15
