+++ soql-server-pg/docker/soql-server-pg.conf.j2
@@ -7,6 +7,11 @@ com.socrata.soql-server-pg {
   port = 6090
   liveness-check.port = 6091
 
+  cache {
+    max-cached-results: {{ PG_SECONDARY_MAX_CACHED_RESULTS }}
+    max-result-size: {{ PG_SECONDARY_MAX_CACHED_RESULT_SIZE }}
+  }
+
   metrics {
     {% if SOQL_SERVER_METRICS_PREFIX is defined -%}
     prefix = "{{ SOQL_SERVER_METRICS_PREFIX }}"
