+++ soql-server-pg/src/main/scala/com/socrata/pg/server/config/QueryServerCacheConfig.scala
@@ -1,0 +1,9 @@
+package com.socrata.pg.server.config
+
+import com.typesafe.config.Config
+import com.socrata.thirdparty.typesafeconfig.ConfigClass
+
+class QueryServerCacheConfig(config: Config, root: String) extends ConfigClass(config, root) {
+  val maxCachedResults = getInt("max-cached-results")
+  val maxResultSize = getInt("max-result-size")
+}
