+++ soql-server-pg/src/main/scala/com/socrata/pg/server/config/QueryServerConfig.scala
@@ -20,4 +20,5 @@ class QueryServerConfig(val config: Config, val root: String) extends ConfigClas
   val maxConcurrentRequestsPerDataset = getInt("max-concurrent-requests-per-dataset")
   val leadingSearch = getBoolean("leading-search")
   val httpQueryTimeoutDelta = getDuration("http-query-timeout-delta")
+  val cache = new QueryServerCacheConfig(config, path("cache"))
 }
