+++ soql-server-pg/src/test/scala/com/socrata/pg/server/QueryServerTest.scala
@@ -14,7 +14,7 @@ import scala.concurrent.duration.Duration
  * @param pgu
  */
 class QueryServerTest(dsInfo:DSInfo, pgu: PGSecondaryUniverse[SoQLType, SoQLValue]) extends
-  QueryServer(dsInfo, CaseSensitive, leadingSearch = true, Duration.Zero) {
+  QueryServer(dsInfo, CaseSensitive, leadingSearch = true, Duration.Zero, new analyzer2.ProcessQuery(analyzer2.NoopResultCache)) {
 
   override protected def withPgu[T](dsInfo:DSInfo, truthStoreDatasetInfo:Option[DatasetInfo])(f: (PGSecondaryUniverse[SoQLType, SoQLValue]) => T): T = {
     f(pgu)
